moving functions around and reorganizing solving equations
